requirements.txt
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
matplotlib==3.8.2
scikit-learn==1.4.0